+++ client/src/components/MajorComponents/DC.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Row, Col, Container } from "reactstrap";
-import '../styles/Major.css';
+import '../../styles/Major.css';
 import ClassDataTooltips from "./ClassDataTooltips";
 
 function parseTakenClasses(taken, classes) {
